Добавил в класс Main3 метод вывода краткого имени компании и даты основания в отформатированном виде
